Remove tensorflow 1.7 dependency
The package has been upgraded to tensorflow 1.14, because a security
warning has been issued for the old tensorflow package. Additionally
I've decided to remove tensorflow as a hard requirement altogether and
let the user decide which version to install or install the recommended
version via extra_requirements.

Tested under linux with tensorflow 1.14 and tensorflow-gpu 1.14