Setup make file for build & testing
To have a consistent build- and testing script I've created a Makefile.
This should help in easily generating meta data like the version string
without relying on more complex frameworks such as versioneer. Also
updated the travis ci yml so it uses the make file, to keep it things
consistent.