WIP: Add first learning tests
This is the start of implementing learning tests which should check on
trivial environments if an RL algorithm is learning something at all.
Additionally it forces me to create interfaces and modules for the other
components necessary for a rl algorithm to work.

Currently in the process of implementing it, but it is uploaded so I've
access to it on all my devices.