Add basic agent implementation & interface
To keep make it easier keeping different agents consistent an Agent
interface has been introduced. This interface acts as a contract that
each agent must fulfill to be used with the algorithms of this library.
Additionally a first batch agents has been implemented, which will be
used by the reinforce algorithm and the first tests.