Refactor trajectory recording
Cleaned up the piecewise recording of trajectories so it clearer how the
interface is supposed to be used. Also moved it into its own class so
there is a clear distinction between the different methods of creating
trajectories. Additionally, the TrajectoryBuilder has been moved to
the testing module as it is solely used by the automated tests.