Implement basic learning test
Add a test that checks if an RL algorithm actually learns something on a
fairly simple environment. However, at the moment it is still WIP,
because the agent doesn't seem to be able to consistently beat the
random agent. There might still be a bug in the algorithm or test and
further investigation is required.