Implement first RL learning test
These automated tests using a most simple environment make sure that the
REINFORCE algorithm is actually learning something and converging
properly. They can also be used for fast diagnosis and understanding of
how the mechanics of RL work. Because the environment is so simple, a
fast iteration cycle and step by step debugging is possible. Still it
can shed light on many issues typically encountered when designing RL
algorithm (i.e. normalization errors). It also gives a feel on what
makes an algorithm stable and what not.