Test baseline propagation & normalization
For the reinforce algorithm to converge it is important, that the value
estimates of the baseline also contain the accumulated reward which the
agent is using as a signal. Otherwise, the baseline adjustment is too
much off to be useful. Hence we introduced the one dimensional walk
environment, where the reward needs to propagate back to the start
position for the agent to stably learn the optimal policy. Additionally
the environment tests, whether the baseline estimates are done on
trajectory batches. This is important because one edge case is where the
basline would estimate perfect values for each state, reducing the
advanatage signals to almost 0 resulting in no useful signal at all.