Add negative test for baseline usage
This test should make sure, that the environment we test the baseline
with is truly harder without usage of a baseline to reduce variance.
While creating the test, it also turned out that the baseline seems to
work much better without normalization of the returns. I suspect this
results from moving the state values by normalizing it. However, by
just normalizing the advantages, the difference between using a baseline
or not, is much clearer.

The tests are a bit brittle right now, however, they should warrant
further investigation of one of them fails.