Add stochastic test marker
To make the reinforce and RL tests more reliable, a stochastic test
marker has been added, which provides a mechanism to record multiple
test runs and assert them on stochastic analysis. In this case baseline
improvement is now measured over multiple runs and can be asserted
statistically (i.e. that in most cases using a baseline is beneficial
to learning, respectively not using it hinders it in most cases.